# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IBN News Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ArticleStatus = typeof ArticleStatus[keyof typeof ArticleStatus];


export const ArticleStatus = {
  draft: 'draft',
  pending_review: 'pending_review',
  published: 'published',
} as const;

export interface Article {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  category: string;
  content: string;
  author: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  isFeatured: boolean;
  isBreaking: boolean;
  viewCount: number;
  status: ArticleStatus;
  publishedAt: string;
  createdAt?: string;
}

export interface ArticleInput {
  /** @minLength 1 */
  title: string;
  subtitle?: string;
  category: string;
  /** @minLength 1 */
  content: string;
  author: string;
  imageUrl?: string;
  videoUrl?: string;
  isFeatured?: boolean;
  isBreaking?: boolean;
}

export interface ArticleUpdate {
  title?: string;
  subtitle?: string;
  category?: string;
  content?: string;
  author?: string;
  imageUrl?: string;
  videoUrl?: string;
  isFeatured?: boolean;
  isBreaking?: boolean;
}

export type ArticlesSummaryByCategoryItem = {
  category: string;
  count: number;
};

export interface ArticlesSummary {
  total: number;
  breaking: number;
  featured: number;
  byCategory: ArticlesSummaryByCategoryItem[];
}

export interface SearchResult {
  results: Article[];
  total: number;
  query: string;
  category: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface SignUpInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  role: string;
}

export interface AuthResponse {
  user: AdminUser;
}

export interface EditorUser {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  role: string;
  /** @nullable */
  lastLoginAt?: string | null;
  createdAt: string;
}

export interface CreateEditorInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface UpdateEditorInput {
  /** @minLength 6 */
  password?: string;
}

export type EngagementStatsTopArticlesItem = {
  id: number;
  title: string;
  category: string;
  author: string;
  viewCount: number;
  publishedAt: string;
};

export type EngagementStatsViewsByCategoryItem = {
  category: string;
  views: number;
};

export interface EngagementStats {
  totalViews: number;
  topArticles: EngagementStatsTopArticlesItem[];
  viewsByCategory: EngagementStatsViewsByCategoryItem[];
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface SubscribeInput {
  email: string;
  name?: string;
}

export interface Subscriber {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  active: boolean;
  subscribedAt: string;
}

export interface Comment {
  id: number;
  articleId: number;
  name: string;
  /** @nullable */
  email?: string | null;
  content: string;
  approved: boolean;
  createdAt: string;
}

export interface CommentInput {
  name: string;
  email?: string;
  /** @maxLength 2000 */
  content: string;
}

export type ListArticlesParams = {
category?: string;
limit?: number;
offset?: number;
};

export type SearchArticlesParams = {
q?: string;
category?: string;
limit?: number;
offset?: number;
};

export type GetTrendingArticlesParams = {
limit?: number;
};

export type PostComment201 = {
  message: string;
  id: number;
};

export type Subscribe200 = {
  message: string;
};

export type Subscribe201 = {
  message: string;
  id?: number;
};

export type UnsubscribeBody = {
  email: string;
};

export type GetPendingCount200 = {
  count: number;
};

export type GetPendingCommentsCount200 = {
  count: number;
};