# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IBN News Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List published articles
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
...
  /articles/trending:
    get:
      operationId: getTrendingArticles
      tags: [articles]
      summary: Get most-viewed articles (trending)
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Trending articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
...
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get a single article (also increments view count)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
  /articles/{id}/comments:
    get:
      operationId: getArticleComments
      tags: [comments]
      summary: Get approved comments for an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: postComment
      tags: [comments]
      summary: Submit a comment (pending review)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment submitted
          content:
            application/json:
              schema:
                type: object
                required: [message, id]
                properties:
                  message:
                    type: string
                  id:
                    type: integer
        "400":
          description: Validation error
        "404":
          description: Article not found
...
  /admin/articles:
    get:
      operationId: adminListArticles
      tags: [admin]
      summary: List all articles (admin)
      responses:
        "200":
          description: All articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [admin]
      summary: Create a new article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /admin/articles/{id}:
    get:
      operationId: adminGetArticle
      tags: [admin]
      summary: Get article for editing (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
    patch:
      operationId: updateArticle
      tags: [admin]
      summary: Update an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
    delete:
      operationId: deleteArticle
      tags: [admin]
      summary: Delete an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /admin/articles/{id}/submit:
    post:
      operationId: submitArticleForReview
      tags: [admin]
      summary: Submit article for admin review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          description: Already published
        "404":
          description: Not found
  /admin/articles/{id}/publish:
    post:
      operationId: publishArticle
      tags: [admin]
      summary: Publish an article (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "403":
          description: Admin role required
        "404":
          description: Not found
  /admin/articles/{id}/unpublish:
    post:
      operationId: unpublishArticle
      tags: [admin]
      summary: Unpublish / send back to draft (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "403":
          description: Admin role required
        "404":
          description: Not found
...
  /admin/comments/{id}/approve:
    patch:
      operationId: approveComment
      tags: [admin]
      summary: Approve a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: Not found
  /admin/comments/{id}/reject:
    patch:
      operationId: rejectComment
      tags: [admin]
      summary: Reject a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: Not found
  /admin/comments/{id}:
    delete:
      operationId: deleteComment
      tags: [admin]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Article:
      type: object
      required: [id, title, category, content, author, publishedAt, isFeatured, isBreaking, viewCount, status]
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        category:
          type: string
        content:
          type: string
        author:
          type: string
        imageUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isBreaking:
          type: boolean
        viewCount:
          type: integer
        status:
          type: string
          enum: [draft, pending_review, published]
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ArticleInput:
      type: object
      required: [title, category, content, author]
      properties:
        title:
          type: string
          minLength: 1
        subtitle:
          type: string
        category:
          type: string
        content:
          type: string
          minLength: 1
        author:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        isFeatured:
          type: boolean
        isBreaking:
          type: boolean
    ArticleUpdate:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        category:
          type: string
        content:
          type: string
        author:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        isFeatured:
          type: boolean
        isBreaking:
          type: boolean
...
    Comment:
      type: object
      required: [id, articleId, name, content, approved, createdAt]
      properties:
        id:
          type: integer
        articleId:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        content:
          type: string
        approved:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [name, content]
      properties:
        name:
          type: string
        email:
          type: string
        content:
          type: string
          maxLength: 2000